+++ changes.md
@@ -1,5 +1,8 @@
 Changes
 ===========
 
-Getting Started
+0.8.7
 ------
+*  Deprecated gson requests - these will be moved to their own library in 0.9
+*  Removed access to the FileRequests and FileQueue. These are still used to back the crossbow image loader and you
+can still use the file image loader.